Eliminate explicit handler functions
There is no good reason to have both a dnel--handle-CloseNotification
handler and a directly called function dnel-close-notification. To be
able to use the former (applied partially) as a handler, the order of
arguments as well as the default reason had to be changed.

In the analogous case, for newly opened notifications, only a handler
dnel--handle-Notify existed, as there was no use case for a directly
called function dnel--notify. It turns out to be useful for testing,
however, so simply rename it.