Never return a value from CloseNotification method
Without the :ignore symbol, a boolean false is returned. Clients don't
seem to mind, but it is better to comply with the spec by explicitly
not returning anything.

This was already done in the old handler, but now the symbol :ignore is
returned if the function is called directly, which doesn't hurt either.