Only signal error where required by specification
The function dnel-get-notification is the easiest way to determine
whether a notification exists. If it doesn't, signalling an error
that the caller must subsequently ignore seems wrong. Instead just
return nil and, if necessary, let the caller signal an error.

(That only happens in response to CloseNotification, for now. It could
also make sense for Notify with an invalid Replaces ID. The spec makes
no mention of that case, so for now a new sequential ID is chosen.)