Improve efficiency of example code for use case 3
The log contains all past and present notifications. The log is updated
on appending a new active notification, or on closing or replacing any
of the active notifications. At no point is anything done to previously
closed notifications, so if an entry's face inherits from shadow, skip
it without even parsing its ID (much less looking it up).

The complexity is still a problem if many notifications are active at
the same moment, but that should be a much less common case than simply
having a long-running log.