Implement a proof-of-concept for image support
Images, as defined by the specification for Desktop Notifications, are
converted to image descriptors (of type PBM) as soon as possible. For
image paths, this is a straightforward call to create-image.

The more complicated case is the one where image data is received on
the D-Bus: ideally it should be enough to simply replace the Desktop
Notifications header with a PBM (or, more correctly, a PNM) header,
without even looking at the raw data. The raw data, however, might
contain an alpha channel, and Emacs does not yet support PNM formats
with transparency. For this reason, the alpha channel must first be
discarded.

(Doing alpha blending here instead of just discarding the channel is
not recommended, as the real background color is still unknown. Brief
experiments in Lisp also introduced a prohibitive performance hit.)