Eliminate additional copy of image data by format
Only the ASCII PNM header needs formatting. Appending it as a list to
the image data (also a list) is fine, since the header is so short and
the last argument is only used as the result's tail. In the end, the
unibyte-string function still creates a copy, but making one copy is
better than making two of them.