Handle images with non-zero padding
Non-zero padding seems to be exceedingly rare in the wild, but a fully
compliant server is still supposed to remove it. As explained in the
commit message of 5104224, the alpha channel has to be removed as well.
This is actually a special case of padding removal, so call a single
function for both of those operations.

A most general implementation of dnel--delete-padding is unfortunately
quite a bit slower than the dnel--delete-every-fourth function that was
used for alpha channel removal before: (partial application of) nthcdr
in the former is a regular function, whereas the cdd(d)r in the latter
is a macro that compiles to repeated uses of the opcode for cdr.

Since alpha channels have to be removed all the time, and other padding
has to be removed almost never, it is justifiable to add some code to
simply resort to the faster macros in the special case.

This optimization should be removed once Emacs gains the native ability
to read PNM images with transparency.