Keep a notification log in *dnel-log* by default
Certain capabilities depend on being able to display data in a proper
buffer (i.e., not just in the mode-line). Logging, use case 3 in the
README, already inserts notifications into a regular buffer, so make
this functionality available by default.

Performance considerations suggest that the code from the README should
not simply be imported into the package: a known problem, for example,
are long lived logs with many old lines that have to be iterated over.
With a normal hook, the update function has no way of knowing which
notification was removed, so replace it by an abnormal hook. This
enables the update function to keep track of log positions in the
notifications themselves and completely eliminates the need for
iterating over lines.

(Since the update function is now part of the package, an abnormal hook
is not strictly necessary: the update function could be called directly
instead, before or after running the normal hook. While closer to the
letter of the recommendations in the Emacs Lisp Reference Manual, it
would prevent users from implementing alternative logging mechanism
without taking a performance hit.)