Never overwrite notification in log
This is essentially a rewrite of dnel--update-log, with the invariant
above as its most important consequence. Not overwriting notifications
(not even when the new notification replaces an older one with the same
ID) seems to be an appropriate behaviour when keeping a log. Crucially,
it also makes tracking of log positions a lot easer, because this way
new notifications are always appended: a recorded log position is thus
always valid for at least as long as the buffer exists.