Optionally append body to formatted notification
The result of the dnel-format-notification function is used both in
places where only a single line is available (e.g., in the mode-line),
as well as in places without any limitations of this kind (e.g., in a
log buffer). The latter situation can now be signalled by an optional
flag, so that the body (potentially consisting of multiple lines) is
appended to the returned string.