Modify controls according to the new format flag
In particular, change actions associated with middle click for a single
line representation of the notification: instead of opening a menu with
the actions, pop to the notification in the log buffer. The log buffer
uses a representation with multiple lines, and pressing the middle
mouse button there still opens a menu, just as before.

The rationale is that short representations might omit some important
data (e.g., the body), so it should be as painless as possible to get
from there to wherever this information is available.

Since dnel-format-notification appends the body to the returned string,
there is no longer any need for dnel--format-summary to set the body as
a help-text. Remove that code and also the now obsolete parameter.