Take notification instead of ID wherever possible
This makes it harder to misuse the API: every notification is typically
a member of dnel-state, since neither constructor nor copier is part of
the public API. An ID, in contrast, is just some integer that need not
map to an existing notification, a situation that can no longer be
expressed in an API where all functions take notifications directly.
(Consequently, also remove some tests that involve unused IDs.)

A notable exception to this new principle is the CloseNotification
handler, which (obviously) still takes an ID. Here it makes sense to
reintroduce a separate (private) handler function that in turn searches
the state for a notification with the given ID and resorts to a public
dnel-close-notification functions to do the rest.

The latter function comes with its own complications, however, because
as opposed to many other functions (e.g., dnel-invoke-action) it still
needs access to the state for removing the notification. For now it is
given as an argument, but this will soon be improved.