Introduce functions to push and pop notifications
The remove-from slot of a notification now always references the cons
cell in the state that has the notification in question as its cdr. The
state object is essentially a doubly linked list where insertions only
happen at the second position, whereas deletions happen at the second
position of an arbitrary suffix of the state. The new push and pop
functions handle linking of the list in O(1).