Make push and pop functions more conventional
It is not immediately obvious what a state suffix is, and for the push
function no non-trivial suffix is used, so just take a state there. For
the pop function, take a notification to extract the suffix from in the
body. The function can hardly be identified as a pop operation now, so
rename it.

The exact meaning of a state suffix is not immediately obvious,