Eliminate the dnel--get-notification function
Since dnel--delete-notification now takes a notification instead of a
suffix, a specialized function to find the suffix is unnecessary and
can be replaced by one that simply finds the element. Such a function,
of course, is already available as cl-find. (IDs are always integers,
so use eq for comparison instead of the default eql.)

This commit also reverts an inadvertent change of behaviour introduced
by commit beb4d65: the specification is silent on what should happen if
an unknown Replaces ID is specified, but simply choosing a new ID is
probably better than rejecting the notification altogether.

A new test ensures that this newly restored behaviour is here to stay,
and the old tests for dnel--get-notification test are no longer needed.