Use the special variable for state directly
So far almost nothing refers to the dnel-state object directly, only
indirectly through the various state parameters that ultimately come
from either dnel-mode or from the user. Looking at other Emacs Lisp
projects (e.g., Elfeed) shows that this is very unusual. Consequently,
any advantages gained from the avoidance of dynamic scope are probably
offset by the confusion caused to the user of the API.

Removing the state parameters means that there is no need to access
that variable from outside of the package, so rename it to dnel--state.
Further rename the hook to dnel-notifications-changed-functions, since
the remaining state is no longer of interest outside of the package.