Add a proper variable definition for the hook
The dnel--update-log-buffer function should never be removed on server
shutdown (functions added by the user are not removed either, and after
a server restart those are suddenly in a different position relative to
the newly pushed dnel--update-log-buffer). Set it as the initial value
instead.

For the tests, rebind the hook variable to nil to ensure that no
unexpected function calls take place. Log tests call the update
function explicitly, so those still work.