Reorganize README
Keep the Commentary as it is, but split the rest of the README into two
sections: Basic Usage, which describes what DNel can do out of the box,
and Advanced Usage, which requires some knowledge of Emacs Lisp.

The two remaining use cases are not different enough to deserve their
own sections. This newly unified section might get a sibling later on,
since nothing in the README uses the hook's arguments yet.