Make some functions interactive
Both dnel-invoke-action and dnel-close-notification take a notification
as their first argument. Make these functions interactive by operating
on the notification referred to by the newly added text properties of
the string at point (see commit 482efd2). This is especially useful for
binding them to keys in the (entire!) log buffer.

Unfortunately, the closures in the old keymap for mouse controls cannot
be replaced by the new commands, as clicking on a formatted string does
not does not necessarily set point to it.