Change arguments of hook to two notifications
When updating the log buffer, it is incorrect to decide on the basis of
the new notification's log position whether it replaces an older one:
its log position is always nil at that point. Copying the log position
to the new notification in dnel--notify would be an easy fix, but it is
better to make replacements more explicit by changing the arguments of
the hook functions.

According to the specification, replacements should happen "with no
flicker". This implies that removal of a replaced notification and
addition of a replacing notification should be signalled within one
single call. It is useful to be able to access both an old removed
notification as well as a newly added one (e.g., in dnel--update-log),
so simply use those (or nil, if not applicable) as the hook arguments.

Add a test to ensure correct formatting of replaced notifications, and
also ensure identical formatting of notifications closed the normal way
(i.e., in the previous test).