Add a command/keybinding to toggle body visibility
As opposed to the other commands, this one operates on (a position in)
a formatted string instead of on a notification. A single notification
might appear in different buffers, after all, with its body visible in
a proper subset of those buffers only.

So far, the optional flag to dnel-format-notifications determined the
visibility of bodies. Another consequence of a non-nil value for that
flag was that pressing the middle mouse button opens the action menu.
Use C-mouse-1 for dnel-pop-to-log-buffer, so that down-mouse-2 opens
that menu unconditionally. Body visibility is taken care of by the
newly added command, so remove the now obsolete flag.

When popping to the log buffer and selecting a specific notification,
the user is presumably interested in details, so expand the body.