Move all data set by hooks into a hook-data alist
Users with their own hooks are free to push data to that alist as well.
The Emacs manual advises that it should not matter in which order hooks
are executed, but for the ednc--update-log-buffer function it clearly
does: it needs to access data set by other hooks (e.g., images and
controls), after all.