Split hook functions into two lists
As mentioned in the message of commit 40342ad, hooks should not depend
on the order in which they are executed, but some of the default hooks
extract data from notifications, and other default hooks take that data
into account when displaying the notification. Users would have to be
careful to add their hooks at the right position.

To prevent troubles like this, replace the old single hook by two new
hooks: one for amending data to notifications (which runs first), the
other for presenting the notification to the user (which runs second).
There is probably no good reason to ever amend data to a notification
that is being removed, so only run the former hook for newly added
notifications (and with only one argument, obviously).