Notify in tests by calling notifications-notify
It is better to rely on a library function in the tests, as it avoids
situations where tests fail because of bugs orthogonal to the function
that is being tested. Of more immediate interest, however, is that the
notifications-notify function takes callback arguments that can be used
to test (for example) whether ednc--invoke-action invokes the correct
action.

The notifictions-notify function takes a property list as its argument,
instead of arguments in a specific order. This simplifies overriding
arguments (so much, in fact, that it no longer makes send to test it),
but required a rewrite of the ednc--test-args-match helper.