Only accept absolute image paths (schema optional)
According to the specification, a URI with the appropriate schema is a
path and everything else is an icon theme name. Unfortunately, this is
not true in reality: some applications forget to prefix paths with the
schema. There is also no good reason to have relative paths, so accept
absolute paths (i.e., strings starting with a slash) only, regardless
of whether they are prefixed or not.

In addition, simply ignore empty paths/icon theme names for now.