Insert file contents with mm-insert-file-contents
Both insert-file-contents-literally and mm-insert-file-contents seem
to have been written with roughly the same goal in mind, with the
latter having the advantage of inhibiting the file handlers in
mm-inhibit-file-name-handlers.

This way, interference by openwith can be avoided by following its
documented suggestion for preventing interference with Gnus:
> (add-to-list 'mm-inhibit-file-name-handlers 'openwith-file-handler)

I hope this fixes issue #9 for GitHub user @NapoleonWils0n.
Thanks for the report!