Scale icon according to line height in frame
Specifically, disregard line height at point: this only made sense if
point was in the log buffer, and even then it had problems when the
notification was shown elsewhere (e.g., in the mode line).

The situation is still not ideal, as there's no good solution for what
size to choose when displaying it in two ore more places at the same
time, but at least icons can be scaled manually since commit bf17d22.